simple-requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain
langchain-openai
python-dotenv
httpx
